refactor(attachment): remove namespace Helli
